Share router and routable
