--- Sources/iOS/Routable.swift
@@ -1,6 +1,0 @@
-import UIKit
-
-public protocol Routable {
-
-  func resolve(arguments: [String: String], fragments: [String : AnyObject], currentController: UIViewController)
-}
